refactor: move self supported protocols check to separate module
